+++ src/Aks/Aks.Test/ScenarioTests/NodePoolTests.ps1
@@ -58,3 +58,94 @@ function Test-NewNodePool
         Remove-AzResourceGroup -Name $resourceGroupName -Force
     }
 }
+
+function Test-NodePoolMode
+{
+    # Setup
+    $resourceGroupName = Get-RandomResourceGroupName
+    $kubeClusterName = Get-RandomClusterName
+    $location = 'eastus'
+    $nodeVmSize = "Standard_D2_v2"
+
+    try
+    {
+        New-AzResourceGroup -Name $resourceGroupName -Location $location
+        
+        # creat default pool, mode=system
+        New-AzAksCluster -ResourceGroupName $resourceGroupName -Name $kubeClusterName -NodeVmSize $nodeVmSize -NodeCount 1
+        
+        $cluster = Get-AzAksCluster -ResourceGroupName $resourceGroupName -Name $kubeClusterName        
+        Assert-AreEqual 1 $cluster.AgentPoolProfiles.Count
+        Assert-AreEqual "default" $cluster.AgentPoolProfiles[0].Name
+        Assert-AreEqual "System" $cluster.AgentPoolProfiles[0].Mode
+        
+        $pools = Get-AzAksNodePool -ResourceGroupName $resourceGroupName -ClusterName $kubeClusterName
+        Assert-AreEqual "default" $pools.Name
+        Assert-AreEqual "System" $pools.Mode
+
+        
+        # create the 2nd nodepool, default mode, mode=User
+        New-AzAksNodePool -ResourceGroupName $resourceGroupName -ClusterName $kubeClusterName -Name "pool2" -Count 1
+        
+        $cluster = Get-AzAksCluster -ResourceGroupName $resourceGroupName -Name $kubeClusterName
+        Assert-AreEqual 2 $cluster.AgentPoolProfiles.Count
+        Assert-AreEqual "System" ($cluster.AgentPoolProfiles | where {$_.Name -eq "default"}).Mode
+        Assert-AreEqual "User" ($cluster.AgentPoolProfiles | where {$_.Name -eq "pool2"}).Mode
+        
+        $pools = Get-AzAksNodePool -ResourceGroupName $resourceGroupName -ClusterName $kubeClusterName
+        Assert-AreEqual 2 $pools.Count
+        Assert-AreEqual "System" ($pools | where {$_.Name -eq "default"}).Mode
+        Assert-AreEqual "User" ($pools | where {$_.Name -eq "pool2"}).Mode
+
+        # create the 3rd nodepool, mode=System
+        New-AzAksNodePool -ResourceGroupName $resourceGroupName -ClusterName $kubeClusterName -Name "pool3" -Count 1 -Mode System
+
+        $cluster = Get-AzAksCluster -ResourceGroupName $resourceGroupName -Name $kubeClusterName
+        Assert-AreEqual 3 $cluster.AgentPoolProfiles.Count
+        Assert-AreEqual "System" ($cluster.AgentPoolProfiles | where {$_.Name -eq "default"}).Mode
+        Assert-AreEqual "User" ($cluster.AgentPoolProfiles | where {$_.Name -eq "pool2"}).Mode
+        Assert-AreEqual "System" ($cluster.AgentPoolProfiles | where {$_.Name -eq "pool3"}).Mode
+
+        $pools = Get-AzAksNodePool -ResourceGroupName $resourceGroupName -ClusterName $kubeClusterName
+        Assert-AreEqual 3 $pools.Count
+        Assert-AreEqual "System" ($pools | where {$_.Name -eq "default"}).Mode
+        Assert-AreEqual "User" ($pools | where {$_.Name -eq "pool2"}).Mode
+        Assert-AreEqual "System" ($pools | where {$_.Name -eq "pool3"}).Mode
+
+        # update the 3rd nodepool, mode=User
+        Update-AzAksNodePool -ResourceGroupName $resourceGroupName -ClusterName $kubeClusterName -Name "pool3" -Mode User
+
+        $cluster = Get-AzAksCluster -ResourceGroupName $resourceGroupName -Name $kubeClusterName
+        Assert-AreEqual 3 $cluster.AgentPoolProfiles.Count
+        Assert-AreEqual "System" ($cluster.AgentPoolProfiles | where {$_.Name -eq "default"}).Mode
+        Assert-AreEqual "User" ($cluster.AgentPoolProfiles | where {$_.Name -eq "pool2"}).Mode
+        Assert-AreEqual "User" ($cluster.AgentPoolProfiles | where {$_.Name -eq "pool3"}).Mode
+
+        $pools = Get-AzAksNodePool -ResourceGroupName $resourceGroupName -ClusterName $kubeClusterName
+        Assert-AreEqual 3 $pools.Count
+        Assert-AreEqual "System" ($pools | where {$_.Name -eq "default"}).Mode
+        Assert-AreEqual "User" ($pools | where {$_.Name -eq "pool2"}).Mode
+        Assert-AreEqual "User" ($pools | where {$_.Name -eq "pool3"}).Mode
+
+        # update the 2nd nodepool, mode=System
+        Set-AzAksCluster -ResourceGroupName $resourceGroupName -Name $kubeClusterName -NodeName "pool2" -NodePoolMode System
+
+        $cluster = Get-AzAksCluster -ResourceGroupName $resourceGroupName -Name $kubeClusterName
+        Assert-AreEqual 3 $cluster.AgentPoolProfiles.Count
+        Assert-AreEqual "System" ($cluster.AgentPoolProfiles | where {$_.Name -eq "default"}).Mode
+        Assert-AreEqual "System" ($cluster.AgentPoolProfiles | where {$_.Name -eq "pool2"}).Mode
+        Assert-AreEqual "User" ($cluster.AgentPoolProfiles | where {$_.Name -eq "pool3"}).Mode
+
+        $pools = Get-AzAksNodePool -ResourceGroupName $resourceGroupName -ClusterName $kubeClusterName
+        Assert-AreEqual 3 $pools.Count
+        Assert-AreEqual "System" ($pools | where {$_.Name -eq "default"}).Mode
+        Assert-AreEqual "System" ($pools | where {$_.Name -eq "pool2"}).Mode
+        Assert-AreEqual "User" ($pools | where {$_.Name -eq "pool3"}).Mode
+
+        $cluster | Remove-AzAksCluster -Force
+    }
+    finally
+    {
+        Remove-AzResourceGroup -Name $resourceGroupName -Force
+    }
+}